requirements.txt
numpy
pandas
scikit-learn
pillow
pytesseract
sounddevice
soundfile